requirements.txt
asgiref==3.3.1
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
cryptography==3.4.6
defusedxml==0.6.0
Django==3.0.7
django-cleanup==4.0.0
django-crispy-forms==1.8.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==2.8
mysqlclient==2.1.0
oauthlib==3.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pycparser==2.20
PyJWT==2.0.0
PyMySQL==1.0.2
python2==1.2
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.2.0
django-import_export